03 day ready for eval

+++ day03/ex04/ClapTrap.cpp
@@ -150,7 +150,7 @@ void	ClapTrap::takeDamage(unsigned int amount)
 
 void	ClapTrap::beRepaired(unsigned int amount)
 {
-	if (this->_hitPoints + amount >= this->_maxHitPoints)
+	if (this->_hitPoints + (int)amount >= this->_maxHitPoints)
 	{
 		this->_hitPoints = this->_maxHitPoints;
 		std::cout << this->_name << " is fully repaired!" << std::endl;
